# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SI Pesantren",
        "description": "Dokumentasi API untuk Sistem Informasi Pesantren",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Localhost"
        }
    ],
    "paths": {
        "/admin/ppdb-calon-santri/detail/{id}": {
            "get": {
                "tags": [
                    "Calon Santri"
                ],
                "summary": "Detail calon santri",
                "operationId": "9d1f7d86213559afa1a1f4e8c000581e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Calon Santri",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses menampilkan detail calon santri"
                    },
                    "404": {
                        "description": "Calon santri tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/ppdb-calon-santri/cetak-laporan-tahunan": {
            "get": {
                "tags": [
                    "Calon Santri"
                ],
                "summary": "Cetak laporan tahunan calon santri",
                "operationId": "a0022babc517a05559b4e918c18c6564",
                "responses": {
                    "200": {
                        "description": "PDF berhasil dibuat"
                    },
                    "400": {
                        "description": "Data tahun ajaran belum tersedia"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}